+++ src/components/Header.js
@@ -4,6 +4,8 @@ import { useNavigate } from 'react-router-dom';
 import { debounce } from 'lodash';
 import logout from "./Logout";
 import getUserInfo from "./GetUserInfo";
+import IsLogin from "./IsLogin";
+import MyPageIcon from "./MyPageIcon";
 
 const SERVER_URL = "http://localhost:8080";
 
@@ -96,7 +98,6 @@ function Header({ searchTerm, setSearchTerm, autocompleteItems, setAutocompleteI
         navigate('/');
     };
 
-
     return (
         <header className="header">
             <img src={logo} className="logo" alt="kkini logo" onClick={() => navigate('/')}/>
@@ -116,15 +117,11 @@ function Header({ searchTerm, setSearchTerm, autocompleteItems, setAutocompleteI
                 </form>
             </div>
             <div className="nav-icons">
-                <span className="icon" onClick={() => user ? navigate('/user') : navigate('/login')}>👤</span> {/* 마이페이지 아이콘 */}
+                <MyPageIcon user={user} navigate={navigate} />
                 <span className="icon"> ♥ </span> {/* 찜하기 아이콘 */}
             </div>
             <div className="nav-links">
-                {user ? (
-                    <a href="" onClick={(e) => handleLogout(e)}>로그아웃</a>
-                ) : (
-                    <a href="" onClick={() => navigate('/login')}>로그인</a>
-                )}
+                <IsLogin user={user} navigate={navigate} handleLogout={handleLogout} />
                 <a href="#">소개</a>
                 <a href="#">도움말</a>
             </div>
